Split LGraph out to TS file
